+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/Identifiers.kt
@@ -56,7 +56,7 @@ internal object Names {
 
     // Function names
     val REALM_CONFIGURATION_BUILDER_BUILD = Name.identifier("build")
-    val REALM_CONFIGURATION_DEFAULT_CONFIG = Name.identifier("defaultConfig")
+    val REALM_CONFIGURATION_WITH = Name.identifier("with")
     val REALM_OBJECT_INTERNAL_IS_FROZEN = Name.identifier("isFrozen")
     val REALM_OBJECT_INTERNAL_REALML_LIFE_CYCLE = Name.identifier("realmLifeCycle")
     val REALM_OBJECT_INTERNAL_VERSION = Name.identifier("version")