+++ packages/plugin-compiler/src/test/resources/schema/input/Schema.kt
@@ -30,4 +30,4 @@ val conf1 = RealmConfiguration.Builder()
 val conf2 = RealmConfiguration.Builder(schema = setOf(A::class, B::class, C::class))
         .build()
 
-val conf3 = RealmConfiguration.defaultConfig(schema = setOf(A::class, C::class))
+val conf3 = RealmConfiguration.with(schema = setOf(A::class, C::class))