+++ test/src/androidTest/kotlin/io/realm/shared/RealmConfigurationTests.kt
@@ -37,8 +37,8 @@ import kotlin.test.assertTrue
 class RealmConfigurationTests {
 
     @Test
-    fun defaultConfig() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+    fun with() {
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertEquals(
             "${appFilesDirectory()}/${Realm.DEFAULT_FILE_NAME}",
             config.path
@@ -49,7 +49,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultPath() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertEquals(
             "${appFilesDirectory()}/${Realm.DEFAULT_FILE_NAME}",
             config.path
@@ -67,7 +67,7 @@ class RealmConfigurationTests {
     fun path() {
         val realmPath = "HowToGetPlatformPath/default.realm"
 
-        val config = RealmConfiguration.defaultConfig(path = realmPath, schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(path = realmPath, schema = setOf(Sample::class))
         assertEquals(realmPath, config.path)
 
         val configFromBuilder: RealmConfiguration =
@@ -80,7 +80,7 @@ class RealmConfigurationTests {
         val realmPath = "<HowToGetPlatformPath>/custom.realm"
         val realmName = "my.realm"
 
-        val config = RealmConfiguration.defaultConfig(realmPath, realmName, setOf(Sample::class))
+        val config = RealmConfiguration.with(realmPath, realmName, setOf(Sample::class))
         assertEquals(realmPath, config.path)
         // Correct assert: assertEquals("custom.realm", config.name)
         assertEquals("my.realm", config.name) // Current result
@@ -94,7 +94,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultName() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertEquals(Realm.DEFAULT_FILE_NAME, config.name)
         assertTrue(config.path.endsWith(Realm.DEFAULT_FILE_NAME))
 
@@ -108,7 +108,7 @@ class RealmConfigurationTests {
     fun name() {
         val realmName = "my.realm"
 
-        val config = RealmConfiguration.defaultConfig(name = realmName, schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(name = realmName, schema = setOf(Sample::class))
         assertEquals(realmName, config.name)
         assertTrue(config.path.endsWith(realmName))
 
@@ -161,7 +161,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultMaxNumberOfActiveVersions() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertEquals(Long.MAX_VALUE, config.maxNumberOfActiveVersions)
     }
 
@@ -182,7 +182,7 @@ class RealmConfigurationTests {
 
     @Test
     fun notificationDispatcherRealmConfigurationDefault() {
-        val configuration = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val configuration = RealmConfiguration.with(schema = setOf(Sample::class))
         assertTrue(configuration.notificationDispatcher is CoroutineDispatcher)
     }
 
@@ -201,7 +201,7 @@ class RealmConfigurationTests {
 
     @Test
     fun writeDispatcherRealmConfigurationDefault() {
-        val configuration = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val configuration = RealmConfiguration.with(schema = setOf(Sample::class))
         assertTrue(configuration.writeDispatcher is CoroutineDispatcher)
     }
 
@@ -232,7 +232,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultSchemaVersionNumber() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertEquals(0, config.schemaVersion)
     }
 
@@ -244,7 +244,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultDeleteRealmIfMigrationNeeded() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertFalse(config.deleteRealmIfMigrationNeeded)
     }
 
@@ -258,7 +258,7 @@ class RealmConfigurationTests {
 
     @Test
     fun defaultEncryptionKey() {
-        val config = RealmConfiguration.defaultConfig(schema = setOf(Sample::class))
+        val config = RealmConfiguration.with(schema = setOf(Sample::class))
         assertNull(config.encryptionKey)
     }