+++ test/src/androidTest/kotlin/io/realm/shared/SampleTests.kt
@@ -38,7 +38,7 @@ class SampleTests {
     @BeforeTest
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
-        val configuration = RealmConfiguration.defaultConfig(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Sample::class))
+        val configuration = RealmConfiguration.with(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Sample::class))
         realm = Realm.open(configuration)
     }