+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmObjectNotificationsTests.kt
@@ -31,7 +31,7 @@ class RealmObjectNotificationsTests : NotificationTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         configuration =
-            RealmConfiguration.defaultConfig(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
+            RealmConfiguration.with(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
         realm = Realm.open(configuration)
     }