+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmResultsNotificationsTests.kt
@@ -30,7 +30,7 @@ class RealmResultsNotificationsTests : NotificationTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         configuration =
-            RealmConfiguration.defaultConfig(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
+            RealmConfiguration.with(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
         realm = Realm.open(configuration)
     }